Moved rpc implementation to correct package
